requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
djangorestframework==3.15.2
docker==7.1.0
idna==3.7
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
